Refactor push registration service to use FCM token
